Fixed recursion not working

+++ kalk/src/kalk_value/rounding.rs
@@ -156,7 +156,7 @@ fn equivalent_fraction(value: f64) -> Option<String> {
     numer = numer.trunc();
     denom = denom.trunc();
 
-    if denom <= 1f64 || denom >= 100f64 || denom == 10f64 {
+    if denom <= 1f64 || denom >= 100f64 || denom as i64 == 10 {
         return None;
     }
